14.0 [IMP] helpdesk_mgmt: make email communication much more useful
The current email templates and processes have several problems:
o Ticket assignment and stage change emails were addressed to the
  assigned user but sent to the customer.
o Visually inconsistent. Emails sent to the assigned user were
  plain text. Emails sent to the customer were formatted but
  didn't fit in with the visual style of other emails sent from
  Odoo (ie. Sales module).
o When emails are sent there is no link in the email that allows
  the recepient to easily access the ticket.
o When a customer sends the initial email they don't receive
  any feedback that it was received.
o The ticket created from the initial email by a customer
  doesn't have the "Partner Name" or "Partner Email" fields
  set correctly
o Ticket assignment and stage change emails don't include any
  information besides the ticket number; forcing the user to
  open the ticket to find out additional information.

After this commit:
o Ticket assignment and stage change emails are addressed and sent
  to the assigned user.
o All sent emails are visually consistent within the helpdesk and
  Odoo in general.
o All emails include a button to access the ticket.
o When an email initially comes in and it's from a partner a reply
  is sent that says the email was received and is being worked on.
o When an email initially comes in and it's from a partner the
  partner name and email fields are set accordingly.
o When a ticket is created from an email the Channel is
  automatically set to 'Email'.
o In communications to the assigned user there is a bit more
  information in the email than just the ticket number.
  In the ticket assignment:
    - Ticket number
    - Title
    - Partner
    - Priority
    - Description
  In the stage change email:
    - Ticket number
    - Title
    - Partner
    - Priority